Add capability to use logical disk partitions for data disks
A new environment variable is defined for this purpose.  It should
be set to the list of disk partitions to be used.  A unique disk must
be specified for each worker node.  For example,

export DATA_DISK_LIST=sdc1,sdd1,sde1

This example assumes 3 worker nodes.

If the DATA_DISK_LIST is not set, then qcow2 images are allocated
by the scripts.

+++ scripts/helper/add-data-disks.sh
@@ -2,6 +2,10 @@
 
 # Create and attach a data disk to each worker node
 
+# One can attach a disk partition or a qcow2 file.  The qcow2 method is
+# the default, which can be overridden by specifying a comma separated list
+# of disk partitions via the environment variable DATA_DISK_LIST.
+
 set -e
 
 if [ ! -e helper/parameters.sh ]; then
@@ -11,11 +15,28 @@ fi
 
 source helper/parameters.sh
 
+if [ -n "$DATA_DISK_ARRAY" ]; then
+
+	for (( i=0; i<$WORKERS; i++ ))
+	do
+		pname=${DATA_DISK_ARRAY[$i]}
+		echo "Attaching /dev/$pname to test-ocp$OCP_VERSION-worker-$i"
+ 		virsh list | grep worker-$i | tail -n 1 | awk -v var="$i" '{system("virsh attach-disk " $2 " --source /dev/$pname --target vdc --persistent")}'
+	done
+
+	exit 
+fi
+
 # Remember where files were created for virsh_cleanup.sh
 
 echo "$IMAGES_PATH" > ~/.images_path
 
 rm -f $IMAGES_PATH/test-ocp$OCP_VERSION/*.data
+
+if [ "$DATA_DISK_SIZE" -le "0" ]; then
+	echo "No data disks will be created"
+	exit
+fi
 
 for (( i=0; i<$WORKERS; i++ ))
 do
